src/main/java/slightlyspring/imgo/user/repository/UserRepository.java
package slightlyspring.imgo.user.repository;

import org.springframework.stereotype.Repository;
import slightlyspring.imgo.user.domain.User;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class UserRepository {

    @PersistenceContext
    private EntityManager em;

    public Long save(User user) {
        em.persist(user);
        return user.getId();
    }

    public User find(Long id) {
        return em.find(User.class, id);
    }
}
