src/test/java/slightlyspring/imgo/user/repository/UserRepositoryTest.java
package slightlyspring.imgo.user.repository;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.transaction.annotation.Transactional;
import slightlyspring.imgo.user.domain.User;

import static org.assertj.core.api.Assertions.assertThat;

@ExtendWith(SpringExtension.class)
@SpringBootTest
class UserRepositoryTest {

    @Autowired UserRepository userRepository;

    @Test
    @Transactional
    @Rollback(false)
    void testUser() throws Exception {
        User user = new User();
        user.setNickname("userA");

        Long savedId = userRepository.save(user);
        User findUser = userRepository.find(savedId);

        assertThat(findUser.getId()).isEqualTo(user.getId());
        assertThat(findUser.getNickname()).isEqualTo(user.getNickname());
        assertThat(findUser).isEqualTo(user);
        System.out.println("findUser == user" + (findUser == user));

    }
}